aws-generative-ai-for-documents-workshop-assets/stream3_working_with_image/image-outpainting/ui/requirements.txt
boto3
streamlit
requests_auth_aws_sigv4
requests
pillow